server_py/requirements.txt
fastapi==0.122.0
uvicorn==0.22.0
requests==2.31.0
boto3==1.28.39
python-dotenv==1.0.0
google-api-python-client==2.88.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
openai==1.4.0
